create file.js
rebuilding api code and creating file.js to keep common code using to fill up HTML
