+++ src/catalog.html
@@ -7,6 +7,7 @@
     <title>Catalog</title>
     <link rel="stylesheet" href="./sass/index.scss" />
     <script type="module" src="./js/api-catalog-page.js"></script>
+
   </head>
   <body>
     <include src="./partials/header.html"></include>
