+++ src/index.html
@@ -2,12 +2,19 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
+
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+
     <title>Cinemania</title>
+
     <link rel="stylesheet" href="./sass/index.scss" />
+
     <script type="module" src="./js/index.js"></script>
+
   </head>
+  
   <body>
     <include src="./partials/header.html"></include>
     <main>
