+++ src/js/index.js
@@ -7,4 +7,4 @@ import './my-library';
 import './modal';
 import './oops';
 import './buttons-icons';
-import './fetch-api';
+import './api-home';
